Structure refactor: group tools, pep8 names
