--- Luna/interface/buildManager/__init__.py
@@ -1,1 +1,0 @@
-from Luna.interface.buildManager.mainDialog import MainDialog
