--- Luna/interface/configManager/__init__.py
@@ -1,1 +1,0 @@
-from Luna.interface.configManager.mainDialog import MainDialog
