--- Luna/interface/lunaMenu/__init__.py
@@ -1,1 +1,0 @@
-from Luna.interface.lunaMenu.creator import LunaMenu
