+++ Luna/interface/menu/__init__.py
@@ -1,0 +1,1 @@
+from Luna.interface.menu.creator import LunaMenu
