+++ Luna/interface/menu/functions/__init__.py
@@ -1,11 +1,11 @@
 from Luna import Logger
 
 try:
-    from Luna.interface.lunaMenu.functions import mainMenu
+    from Luna.interface.menu.functions import main_menu
 except Exception as e:
     Logger.exception("Failed to import modules", exc_info=e)
 
 try:
-    reload(mainMenu)
+    reload(main_menu)
 except Exception as e:
     Logger.exception("Failed to reload modules", exc_info=e)
