+++ Luna/interface/menu/functions/main_menu.py
@@ -1,8 +1,8 @@
 import pymel.core as pm
 from Luna import Logger
 try:
-    from Luna.interface import buildManager
-    from Luna.interface import configManager
+    from Luna.tools import buildManager
+    from Luna.tools import configManager
 except Exception as e:
     Logger.exception("Failed to import modules", exc_info=e)
