+++ Luna/tools/buildManager/__init__.py
@@ -1,0 +1,1 @@
+from Luna.tools.buildManager.mainDialog import MainDialog
