+++ Luna/tools/configManager/__init__.py
@@ -1,0 +1,1 @@
+from Luna.tools.configManager.mainDialog import MainDialog
